+++ anchorpoint/__init__.py
@@ -1,3 +1,5 @@
+"""Anchorpoint is a library that supplies text substring selectors for anchoring annotations."""
+
 from anchorpoint.textselectors import TextQuoteSelector, TextPositionSelector
 from anchorpoint.textselectors import TextPositionSet, TextPositionSetFactory